Add files via upload
The purpose of this project is to make easy the work of measurements of fundamental characteristics of water bodies such as rivers and lakes. Currently, most measurement routines are performed manually, with low periodicity, since the place where they should be performed is difficult to access. Following this principle, we start with the second point of this work, which is the elaboration of the autonomous system powered by a solar panel battery. The ultimate goal is to integrate measurement systems, controllers, storage, chargers, and transmitters so that they can be accessed remotely, allowing the easy component handling, usability autonomy, and frequent information gathering.